requirements.txt
uwsgi==2.0.19.1
pyyaml==5.3.1